move `makeStep` back to `PathwaysView`
